feat(adapters):add and update alerta-view-model class
also...
feat(adapters):add and update alerta-adapterclass
feat(main):update main-activity class
feat(main):update main-application class
feat(services):update alerta-api class
feat(services):add and update alerta-sismo-api-service class
feat(services): update ultimos-sismos-chile class change float to double
feat(services): update alerta-service class
feat(services): update mockup class
feat(services): add and update transformer class
feat(layout): update activity-main
feat(layout): update row_alertasimsmo

+++ app/src/main/java/cl/ucn/disc/dsm/chilealerta/services/alertapi/UltimosSismosChile.java
@@ -22,11 +22,11 @@ public class UltimosSismosChile {
   public String localTime;
   public String chileanTime;
   public String reference;
-  public Float magnitude;
+  public Double magnitude;
   public String scale;
-  public Float latitude;
-  public Float longitude;
-  public Integer depth;
+  public Double latitude;
+  public Double longitude;
+  public Double depth;
   public String id;
   public String url;
   public String source;
